Bufferable and Model revamp
Moved NIO allocation to helper class.
Added bufferable array wrapper.
Added toByteArray() conversion helper.
Length has to be an integer.
Refactored image to use bufferable.
Moved vertex count back to header and also included layout.
Re-introduced abstract implementation.
Both models now have layout since ALL models will.
TODO - OBJ model names.
Removed superfluous BB load on VBO.
Added clockwise flag throughout.

+++ src/main/java/org/sarge/jove/common/Colour.java
@@ -90,7 +90,7 @@ public record Colour(float red, float green, float blue, float alpha) implements
 	}
 
 	@Override
-	public long length() {
+	public int length() {
 		return SIZE * Float.BYTES;
 	}
